pyproject.toml
[project]
name = "walldesignertool"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "datasets>=4.4.1",
    "matplotlib>=3.10.8",
    "numpy>=2.3.5",
    "opencv-python>=4.11.0.86",
    "pandas>=2.2.0",
    "pillow>=12.0.0",
    "scipy>=1.16.3",
    "torch>=2.9.1",
    "transformers>=4.57.3",
]
